Transferred /create under /api instead of /api/ctf
