+++ src/pwncore/routes/ctf/__init__.py
@@ -5,7 +5,6 @@ from fastapi import Response
 
 from pwncore.models import Problem, SolvedProblem, Container, Hint, ViewedHint, Team
 from pwncore.config import config
-import uuid
 
 # Metadata at the top for instant accessibility
 metadata = {
@@ -16,70 +15,15 @@ metadata = {
 
 router = APIRouter(prefix="/ctf", tags=["ctf"])
 
+
 # Routes that do not need a separate submodule for themselves
-
-
 # Fetch team_id from cookies
 def get_team_id():
     return 1
 
 
-@router.get("/create")
-async def init_db():
-    if config.development:
-        await Problem.create(
-            name="Invisible-Incursion",
-            description="Chod de tujhe se na ho paye",
-            author="Meetesh Saini",
-            points=300,
-            image_name="key:latest",
-            image_config={"PortBindings": {"22/tcp": [{}]}},
-        )
-        await Problem.create(
-            name="In-Plain-Sight",
-            description="A curious image with hidden secrets?",
-            author="KreativeThinker",
-            points=300,
-            image_name="key:latest",
-            image_config={"PortBindings": {"22/tcp": [{}]}},
-        )
-        await Team.create(
-            name="CID Squad" + uuid.uuid4().hex, secret_hash="veryverysecret"
-        )
-        await Team.create(
-            name="Triple A battery" + uuid.uuid4().hex, secret_hash="chotiwali"
-        )
-        await Container.create(
-            docker_id="letsjustsay1",
-            flag="pwncore{this_is_a_test_flag}",
-            problem_id=1,
-            team_id=1,
-        )
-        await Container.create(
-            docker_id="letsjustsay2",
-            flag="pwncore{this_is_a_test_flag}",
-            problem_id=2,
-            team_id=1,
-        )
-        await Container.create(
-            docker_id="letsjustsay3",
-            flag="pwncore{farmers}",
-            problem_id=2,
-            team_id=2,
-        )
-        await Hint.create(order=0, problem_id=1, text="This is the first hint")
-        await Hint.create(order=1, problem_id=1, text="This is the second hint")
-        await Hint.create(order=2, problem_id=1, text="This is the third hint")
-        await Hint.create(order=0, problem_id=2, text="This is the first hint")
-        await Hint.create(order=1, problem_id=2, text="This is the second hint")
-
-
-response = Response()
-
-
 @router.get("/list")
-async def ctf_list():
-    # Get list of ctfs
+async def ctf_list(response: Response):
     problems = await Problem.all().values()
     if not problems:
         response.status_code = 404
@@ -97,7 +41,7 @@ async def ctf_list():
 
 # For testing purposes only. flag to be passed in body of POST function.
 @router.get("/flag/{ctf_id}/{flag}")
-async def flag_get(ctf_id: int, flag: str):
+async def flag_get(ctf_id: int, flag: str, response: Response):
     problem = await Problem.get_or_none(id=ctf_id)
     if not problem:
         response.status_code = 404
@@ -113,7 +57,7 @@ async def flag_get(ctf_id: int, flag: str):
 
 
 @router.get("/hint/{ctf_id}")
-async def hint_get(ctf_id: int):
+async def hint_get(ctf_id: int, response: Response):
     problem = await Problem.get_or_none(id=ctf_id)
     if not problem:
         response.status_code = 404
@@ -163,7 +107,7 @@ async def completed_problem_get():
 
 
 @router.get("/{ctf_id}")
-async def ctf_get(ctf_id: int):
+async def ctf_get(ctf_id: int, response: Response):
     problem = await Problem.get_or_none(id=ctf_id)
     if not problem:
         response.status_code = 404
