+++ src/pwncore/routes/team.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-from fastapi import APIRouter
+from fastapi import APIRouter, Response
+from pwncore.models import Team, User
+from pwncore.config import config
 
 # Metadata at the top for instant accessibility
 metadata = {"name": "team", "description": "Operations with teams"}
@@ -9,9 +11,12 @@ router = APIRouter(prefix="/team", tags=["team"])
 
 
 @router.get("/list")
-async def team_list():
-    # Do login verification here
-    return [{"team_name": "CID Squad"}, {"team_name": "Astra"}]
+async def team_list(response: Response):
+    teams = await Team.all().values()
+    if not teams:
+        response.status_code = 404
+        return {"msg_code": config.msg_codes["team_not_found"]}
+    return teams
 
 
 @router.get("/login")
@@ -20,7 +25,12 @@ async def team_login():
     return {"status": "logged in!"}
 
 
+# Unable to test as adding users returns an error
 @router.get("/members/{team_id}")
-async def team_members(team_id: int):
-    # Get team members from team_id
-    return [{"name": "ABC", "user_id": 3432}, {"name": "DEF", "user_id": 3422}]
+async def team_members(team_id: int, response: Response):
+    team = await Team.get_or_none(id=team_id).values_list("members_id", flat=True)
+    if not team:
+        response.status_code = 404
+        return {"msg_code": config.msg_codes["team_not_found"]}
+    members = [await User.get(id=x) for x in team]
+    return members
